refactor(multiplayer): extract useActiveRoom hook
The setActiveRoomId + roomJoin + cleanup effect was duplicated
verbatim in the room and game pages. Move it into a useActiveRoom
hook and call it from both.
